centralize watchstatus
