--- specification/watcher/ack_watch/WatchStatus.ts
@@ -1,33 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import { Dictionary } from '@spec_utils/Dictionary'
-import { IndexName, VersionNumber } from '@_types/common'
-import { DateString } from '@_types/Time'
-import { ActionStatus } from './ActionStatus'
-import { ActivationState } from './ActivationState'
-
-export class WatchStatus {
-  actions: Dictionary<IndexName, ActionStatus>
-  last_checked?: DateString
-  last_met_condition?: DateString
-  state: ActivationState
-  version: VersionNumber
-  execution_state?: string // TODO find execution states in export enum  in server codebase
-}
