+++ specification/watcher/get_watch/GetWatchResponse.ts
@@ -17,8 +17,7 @@
  * under the License.
  */
 
-import { WatchStatus } from '@watcher/ack_watch/WatchStatus'
-import { Watch } from '@watcher/_types/Watch'
+import { Watch, WatchStatus } from '@watcher/_types/Watch'
 import { Id, SequenceNumber, VersionNumber } from '@_types/common'
 import { integer } from '@_types/Numeric'
